backend/app/services/satellite_verify.py
"""
Satellite Verification Service
Uses before/after satellite imagery to verify if a pothole has been repaired.
In production: integrates with Google Earth Engine, ISRO Bhuvan, or commercial
providers (Maxar, Planet). This module provides the interface + a scoring pipeline.
"""
import logging
import httpx
import numpy as np
from typing import Optional
from datetime import datetime

logger = logging.getLogger(__name__)


class SatelliteVerificationResult:
    def __init__(self, repaired: bool, confidence: float, method: str, notes: str = ""):
        self.repaired = repaired
        self.confidence = confidence
        self.method = method        # "imagery_diff", "api_response", "fallback_manual"
        self.notes = notes
        self.verified_at = datetime.utcnow()


async def verify_pothole_repair(
    pothole_id: str,
    lat: float,
    lon: float,
    reported_repaired_at: datetime,
) -> SatelliteVerificationResult:
    """
    Main entry point for repair verification.
    Tries multiple sources in order of reliability.
    """
    # 1. Try Google Earth Engine (requires API credentials)
    result = await _try_gee_verification(lat, lon, reported_repaired_at)
    if result:
        return result

    # 2. Try Bhuvan (ISRO open satellite API — free for India)
    result = await _try_bhuvan_verification(lat, lon, reported_repaired_at)
    if result:
        return result

    # 3. Fallback: flag for manual review
    logger.warning(f"Satellite verification unavailable for pothole {pothole_id}. Flagging for manual review.")
    return SatelliteVerificationResult(
        repaired=False,
        confidence=0.0,
        method="fallback_manual",
        notes="Satellite imagery unavailable. Queued for manual field verification.",
    )


async def _try_gee_verification(
    lat: float,
    lon: float,
    repair_date: datetime,
) -> Optional[SatelliteVerificationResult]:
    """
    Google Earth Engine via REST API.
    Compares NDVI/road-surface texture before and after repair date.
    Requires GEE service account credentials in environment.
    """
    from app.config import get_settings
    settings = get_settings()

    gee_key = getattr(settings, "GEE_SERVICE_ACCOUNT_KEY", None)
    if not gee_key:
        return None

    try:
        # Bounding box ~10m around pothole
        bbox = _make_bbox(lat, lon, buffer_meters=10)
        before_date = repair_date.strftime("%Y-%m-%d")

        payload = {
            "expression": _build_gee_expression(bbox, before_date),
            "fileFormat": "PNG",
        }

        async with httpx.AsyncClient(timeout=30) as client:
            # In production: use GEE Python client or Maps API
            # This is the interface pattern
            resp = await client.post(
                "https://earthengine.googleapis.com/v1/projects/sadak-sathi/image:computePixels",
                json=payload,
                headers={"Authorization": f"Bearer {gee_key}"},
            )
            if resp.status_code != 200:
                return None

            score = _parse_gee_response(resp.json())
            return SatelliteVerificationResult(
                repaired=score >= 0.70,
                confidence=score,
                method="imagery_diff",
                notes=f"GEE road surface analysis. Score: {score:.2f}",
            )
    except Exception as e:
        logger.warning(f"GEE verification failed: {e}")
        return None


async def _try_bhuvan_verification(
    lat: float,
    lon: float,
    repair_date: datetime,
) -> Optional[SatelliteVerificationResult]:
    """
    ISRO Bhuvan WMS API (open access, India coverage).
    Fetches road-layer tiles before/after and computes pixel diff.
    """
    try:
        bbox = _make_bbox(lat, lon, buffer_meters=15)
        bbox_str = f"{bbox['minx']},{bbox['miny']},{bbox['maxx']},{bbox['maxy']}"

        base_url = "https://bhuvan-app1.nrsc.gov.in/bhuvan/wms"
        params = {
            "SERVICE": "WMS",
            "VERSION": "1.1.1",
            "REQUEST": "GetMap",
            "LAYERS": "bhuvan:india_road",
            "SRS": "EPSG:4326",
            "BBOX": bbox_str,
            "WIDTH": "64",
            "HEIGHT": "64",
            "FORMAT": "image/png",
        }

        async with httpx.AsyncClient(timeout=20) as client:
            resp = await client.get(base_url, params=params)
            if resp.status_code != 200 or "image" not in resp.headers.get("content-type", ""):
                return None

            score = _analyze_road_tile(resp.content)
            return SatelliteVerificationResult(
                repaired=score >= 0.65,
                confidence=score,
                method="imagery_diff",
                notes=f"Bhuvan WMS analysis. Score: {score:.2f}",
            )
    except Exception as e:
        logger.warning(f"Bhuvan verification failed: {e}")
        return None


def _analyze_road_tile(png_bytes: bytes) -> float:
    """
    Simple road surface analysis:
    - Dark, uniform pixels → intact road surface → repaired
    - Irregular, lighter patches → pothole still present
    Returns confidence score 0–1 that road is repaired.
    """
    try:
        import io
        from PIL import Image

        img = Image.open(io.BytesIO(png_bytes)).convert("L")  # grayscale
        arr = np.array(img, dtype=np.float32)

        # Road surface metrics
        mean_brightness = arr.mean()
        std_brightness = arr.std()

        # Intact road: dark (asphalt ~40-80 brightness), low variance
        # Pothole: lighter region (dirt/water), higher local variance

        # Normalise: lower brightness + lower variance → more likely repaired
        brightness_score = 1.0 - min(mean_brightness / 200.0, 1.0)
        uniformity_score = 1.0 - min(std_brightness / 60.0, 1.0)

        score = (brightness_score * 0.4 + uniformity_score * 0.6)
        return float(np.clip(score, 0.0, 1.0))
    except Exception:
        return 0.5  # neutral if analysis fails


def _make_bbox(lat: float, lon: float, buffer_meters: float) -> dict:
    """Approximate bounding box around a point."""
    deg_per_meter_lat = 1 / 111320
    deg_per_meter_lon = 1 / (111320 * np.cos(np.radians(lat)))

    buf_lat = buffer_meters * deg_per_meter_lat
    buf_lon = buffer_meters * deg_per_meter_lon

    return {
        "minx": lon - buf_lon,
        "miny": lat - buf_lat,
        "maxx": lon + buf_lon,
        "maxy": lat + buf_lat,
    }


def _build_gee_expression(bbox: dict, before_date: str) -> str:
    """Build a GEE expression for road surface change detection."""
    return f"""
    var before = ee.ImageCollection('COPERNICUS/S2_SR')
        .filterBounds(ee.Geometry.BBox({bbox['minx']},{bbox['miny']},{bbox['maxx']},{bbox['maxy']}))
        .filterDate('2023-01-01', '{before_date}')
        .median();
    var after = ee.ImageCollection('COPERNICUS/S2_SR')
        .filterBounds(ee.Geometry.BBox({bbox['minx']},{bbox['miny']},{bbox['maxx']},{bbox['maxy']}))
        .filterDate('{before_date}', ee.Date('{before_date}').advance(30, 'day'))
        .median();
    var diff = after.subtract(before).abs();
    return diff.reduceRegion(ee.Reducer.mean());
    """


def _parse_gee_response(response: dict) -> float:
    """Parse GEE pixel response into a 0-1 repair confidence score."""
    try:
        # Lower spectral change = road surface unchanged = not repaired
        # Higher change in road bands = repair material applied
        bands = response.get("result", {})
        b4 = bands.get("B4", 0)   # Red band — sensitive to road material
        b8 = bands.get("B8", 0)   # NIR — sensitive to material texture
        change = (b4 + b8) / 2.0
        return float(np.clip(change / 500.0, 0.0, 1.0))
    except Exception:
        return 0.5