backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
geoalchemy2==0.15.1
alembic==1.13.1
pydantic-settings==2.2.1
redis==5.0.4
celery==5.4.0
ultralytics==8.2.0       # YOLOv8
Pillow==10.3.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.0             # S3 for images
httpx==0.27.0